Revert rename to extensions
